requirements.txt
Flask
flask-cors
python-dotenv
pyjwt
supabase
cloudinary
pinecone-client
huggingface_hub
langchain
langchain-google-genai
pymupdf
nltk
google-generativeai
httpx
retry
reportlab
beautifulsoup4
markdown
gunicorn